ai-service/requirements.txt
fastapi[standard]
google-genai
langchain
langchain-google-genai
langchain-chroma
langchain-community
pypdf
python-dotenv